Refactor: Migrating mock data for CreateHearing
