+++ src/app/routing/pages/hearings/check-hearing/mock/check-hearing.mock.ts
@@ -1,13 +1,5 @@
-export interface SectionfiedItems{
-    id: string;
-    name: string;
-}
-export interface CaseData {
-    id: string;
-    decision: {
-        options: Array<SectionfiedItems>
-    }
-}
+import { CaseData } from '../../../modules/case';
+
 export const mockCaseData: CaseData = {
     id: '1234',
     decision: {
