+++ src/app/routing/pages/hearings/create-hearing/mock/create-hearing.mock.ts
@@ -1,0 +1,18 @@
+import {CaseDataOther} from '../../../modules/case';
+
+export const mockCaseData: CaseDataOther = {
+        id: '1234',
+        sections: [],
+        details: {
+            fields: [
+                {value: '123'},
+                {value: 'bob v bob'}
+            ]
+        },
+        decision: {
+            options: [
+                {id: 'appeal-decline', name: 'Appeal Declined'}
+            ]
+        }
+
+}
